Add params, refactor, HACK to guess initial infections
